+++ MS_LIBRARY/trunk/src/org/yeastrc/ms/parser/sqtFile/percolator/PercolatorSQTFileReader.java
@@ -266,7 +266,7 @@ public class PercolatorSQTFileReader extends SQTFileReader<PercolatorSearchScan>
         public String getServerName() {
             return scan.getServerName();
         }
-        public BigDecimal getTotalIntensity() {
+        public Double getTotalIntensity() {
             return scan.getTotalIntensity();
         }
